User ID is taken from jwt token in endpoints

+++ utils/db/db_functions.py
@@ -1,20 +1,21 @@
-from models import owner
+from fastapi import HTTPException
+from starlette.status import HTTP_404_NOT_FOUND, HTTP_401_UNAUTHORIZED
 from models.sera import Sera
 from utils.db.db import fetch, execute
 from models.owner import Owner
 
 
-async def db_check_username(user):
+async def db_check_username(username):
     # select * from User -> postgres
     query = """select * from users where username = :username """
-    values = {"username": user.username}
+    values = {"username": username}
     result = await fetch(query, False, values)
     return result
 
 
-async def db_check_jwt_token(username):
-    query = """select * from users where username = :username"""
-    values = {"username": username}
+async def db_check_jwt_token(user_id):
+    query = """select * from users where user_id = :user_id"""
+    values = {"user_id": user_id}
     result = await fetch(query, True, values)
     if result is None:
         return False
@@ -42,7 +43,7 @@ async def db_get_all_sera():
     return sera_all
 
 
-# TODO: get only my sera
+# get only my sera
 async def db_get_my_sera(current_user_id):
     query = """ select * from sera
                 where :current_user = any (owners)"""
@@ -54,7 +55,7 @@ async def db_get_my_sera(current_user_id):
     return sera_all
 
 
-# TODO: get only my sera
+# get only my sera
 async def db_get_sera_by_id(sera_id):
     query = """ select * from sera
                 where sera_id = :sera_id"""
@@ -63,7 +64,7 @@ async def db_get_sera_by_id(sera_id):
     return result
 
 
-# TODO: insert a new sera
+# insert a new sera
 async def db_insert_sera(current_user_id, sera):
     array = []
     array.append(current_user_id)
@@ -75,11 +76,14 @@ async def db_insert_sera(current_user_id, sera):
     await execute(query, False, values)
 
 
-# TODO: update sera owners
+# update sera owners
 async def db_add_owner_to_sera(current_user_id, sera_id):
     find_sera = await db_get_sera_by_id(sera_id)
     if find_sera is None:
-        return "Sera is not found! (404)"  # burada 404 diye yazıyla dönüyorsun ama bu bir anlam ifade etmiyor. Adama HTTP 404 dönmen gerekiyor requestine karşılık. şuan HTTP 200 dönüyor içinde 404 yazıyor bu olmaz.
+        # burada 404 diye yazıyla dönüyorsun ama bu bir anlam ifade etmiyor. Adama HTTP 404 dönmen gerekiyor requestine karşılık. şuan HTTP 200 dönüyor içinde 404 yazıyor bu olmaz.
+        # http status buradan mı çıkmalı?
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND,
+                            detail="Sera is not found!")
     else:
         result = await db_get_sera_by_id(sera_id)
         owners = result["owners"]
@@ -97,30 +101,36 @@ async def db_add_owner_to_sera(current_user_id, sera_id):
 
 # TODO: update sera info.
 
-# TODO: delete sera
+# delete sera
 async def db_delete_sera(sera_id, current_user_id):
     # find all owners of this sera
     result = await db_get_sera_by_id(sera_id)
-    owners = result["owners"]
-    # remove current user from owner list
-    try:
-        owners.remove(current_user_id)
-        # if the last owner deletes, delete sera from table
-        if len(owners) == 0:
-            query = """ delete from sera 
-                        where sera_id = :sera_id """
-            values = {"sera_id": sera_id}
-            result = "Sera is deleted completely!"
-        # if there are other owners, just remove current user from list
-        else:
-            query = """ update sera 
-                        set owners = :owners
-                        where sera_id = :sera_id """
-            values = {"sera_id": sera_id, "owners": owners}
-            result = "Sera is deleted!"
+    if result is not None:
+        owners = result["owners"]
+        # remove current user from owner list
+        try:
+            owners.remove(current_user_id)
+            # if the last owner deletes, delete sera from table
+            if len(owners) == 0:
+                query = """ delete from sera 
+                            where sera_id = :sera_id """
+                values = {"sera_id": sera_id}
+                result = "Sera is deleted completely!"
+            # if there are other owners, just remove current user from list
+            else:
+                query = """ update sera 
+                            set owners = :owners
+                            where sera_id = :sera_id """
+                values = {"sera_id": sera_id, "owners": owners}
+                result = "Sera is deleted!"
 
-        await execute(query, False, values)
-        return result
-    # if user is not found in owners
-    except Exception as e:
-        return "You do not own this sera!"
+            await execute(query, False, values)
+            return result
+        # if user is not found in owners
+        except Exception as e:
+            raise HTTPException(status_code=HTTP_401_UNAUTHORIZED,
+                                detail="You are not the owner!")
+    # if there is no such sera
+    else:
+        raise HTTPException(status_code=HTTP_404_NOT_FOUND,
+                            detail="Sera is not found!")
